Clean up graph tools and add DetailsPopup
